LayoutTests/sputnik/Conformance/07_Lexical_Conventions/7.9_Automatic_Semicolon_Insertion/7.9.2_Examples/S7.9.2_A1_T1-expected.txt

CONSOLE MESSAGE: line 77: SyntaxError: Unexpected number '2'
S7.9.2_A1_T1

PASS Expected parsing failure

TEST COMPLETE
